% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateLVConstraints.R
\name{calculateLVConstraints}
\alias{calculateLVConstraints}
\title{calculateLVConstraints}
\usage{
calculateLVConstraints(results, x_axis = 1, y_axis = 2, constraints = NULL)
}
\arguments{
\item{results}{results (with $lx and $ly) from TExPosition (i.e.,
$TExPosition.Data)}

\item{x_axis}{which component should be on the x axis?}

\item{y_axis}{which component should be on the y axis?}

\item{constraints}{if available, axis constraints for the plots (determines
end points of the plots).}
}
\value{
Returns a list with the following items:\cr \item{$constraints}{axis
constraints for the plots (determines end points of the plots).}
}
\description{
Calculates constraints for plotting latent variables.
}
\author{
Derek Beaton
}
\keyword{misc}
