% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.tepAssign.R
\name{print.tepAssign}
\alias{print.tepAssign}
\title{Print assignment results}
\usage{
\method{print}{tepAssign}(x, ...)
}
\arguments{
\item{x}{an list that contains items to make into the tepAssign class.}

\item{\dots}{inherited/passed arguments for S3 print method(s).}
}
\description{
Print assignment results.
}
\author{
Derek Beaton, Cherise Chin-Fatt
}
\keyword{print}
