% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.tepBADA.R
\name{print.tepBADA}
\alias{print.tepBADA}
\title{Print tepBADA results}
\usage{
\method{print}{tepBADA}(x, ...)
}
\arguments{
\item{x}{an list that contains items to make into the tepBADA class.}

\item{\dots}{inherited/passed arguments for S3 print method(s).}
}
\description{
Print tepBADA results.
}
\author{
Derek Beaton, Cherise Chin-Fatt
}
\keyword{print}
