% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.tepDICA.R
\name{print.tepDICA}
\alias{print.tepDICA}
\title{Print tepDICA results}
\usage{
\method{print}{tepDICA}(x, ...)
}
\arguments{
\item{x}{an list that contains items to make into the tepDICA class.}

\item{\dots}{inherited/passed arguments for S3 print method(s).}
}
\description{
Print tepDICA results.
}
\author{
Derek Beaton, Cherise Chin-Fatt
}
\keyword{print}
