\name{print.tepGPLS}
\alias{print.tepGPLS}
\title{Print tepGPLS results}
\description{
	Print tepGPLS results.
}
\usage{
	\method{print}{tepGPLS}(x,\dots)
}
\arguments{
  \item{x}{an list that contains items to make into the tepGPLS class.}
  \item{\dots}{inherited/passed arguments for S3 print method(s).}  
 }
\author{Derek Beaton, Cherise Chin-Fatt}
\keyword{print}
