% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.tepGraphs.R
\name{print.tepGraphs}
\alias{print.tepGraphs}
\title{Print tepGraphs results}
\usage{
\method{print}{tepGraphs}(x, ...)
}
\arguments{
\item{x}{an list that contains items to make into the tepGraphs class.}

\item{\dots}{inherited/passed arguments for S3 print method(s).}
}
\description{
Print tepGraphs results.
}
\author{
Derek Beaton, Cherise Chin-Fatt
}
\keyword{print}
