% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.tepPLS.R
\name{print.tepPLS}
\alias{print.tepPLS}
\title{Print tepPLS results}
\usage{
\method{print}{tepPLS}(x, ...)
}
\arguments{
\item{x}{an list that contains items to make into the tepPLS class.}

\item{\dots}{inherited/passed arguments for S3 print method(s).}
}
\description{
Print tepPLS results.
}
\author{
Derek Beaton, Cherise Chin-Fatt
}
\keyword{print}
