\name{print.tepPLSCA}
\alias{print.tepPLSCA}
\title{Print tepPLSCA results}
\description{
	Print tepPLSCA results.
}
\usage{
	\method{print}{tepPLSCA}(x,\dots)
}
\arguments{
  \item{x}{an list that contains items to make into the tepPLSCA class.}
  \item{\dots}{inherited/passed arguments for S3 print method(s).}  
 }
\author{Derek Beaton, Cherise Chin-Fatt}
\keyword{print}
