\name{print.texpoOutput}
\alias{print.texpoOutput}
\title{Print TExPosition results}
\description{
	Print TExPosition results.
}
\usage{
	\method{print}{texpoOutput}(x,\dots)
}
\arguments{
  \item{x}{an list that contains items to make into the texpoOutput class.}
  \item{\dots}{inherited/passed arguments for S3 print method(s).}  
 }
\author{Derek Beaton, Cherise Chin-Fatt}
\keyword{print}
