% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2.R
\name{R2}
\alias{R2}
\title{R-squared computations}
\usage{
R2(group.masses, di, ind.masses = NULL, dii)
}
\arguments{
\item{group.masses}{a masses matrix for the groups}

\item{di}{a set of squared distances of the groups}

\item{ind.masses}{a masses matrix for the individuals}

\item{dii}{a set of squared distances for the individuals}
}
\value{
\item{R2}{An R-squared}
}
\description{
A function to compute R-squared for BADA and DICA
}
\author{
Jenny Rieck, Derek Beaton
}
\keyword{misc}
\keyword{multivariate}
