\name{TExPosition-package}
\alias{TExPosition-package}
\alias{TExPosition}
\encoding{UTF-8}
\docType{package}
\title{
TExPosition: Two-table analyses with via ExPosition. 
}
\description{
TExPosition is two-table \code{\link{ExPosition}} and includes discriminant methods of the singular value decomposition (SVD). The core of TExPosition is \code{\link{ExPosition}} and the \code{\link{svd}}.
}
\details{
\tabular{ll}{
Package: \tab TExPosition\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-06-13\cr
Depends: \tab R (>=2.15.0), prettyGraphs, ExPosition\cr
License: \tab GPL-2\cr
URL: \tab http://www.utdallas.edu/~derekbeaton/software/ExPosition\cr
}
}
\author{
Questions, comments, compliments, and complaints go to Derek Beaton \email{exposition.software@gmail.com}.\cr\cr

The following people are authors or contributors to TExPosition code, data, or examples:\cr
Derek Beaton, Jenny Rieck, Cherise Chin-Fatt and Hervé Abdi.
}
\references{
Abdi, H., and Williams, L.J. (2010). Principal component analysis. \emph{Wiley Interdisciplinary Reviews: Computational Statistics}, 2, 433-459.\cr
Abdi, H. and Williams, L.J. (2010). Correspondence analysis. In N.J. Salkind, D.M., Dougherty, & B. Frey (Eds.): \emph{Encyclopedia of Research Design}. Thousand Oaks (CA): Sage. pp. 267-278.\cr
Abdi, H. (2007). Singular Value Decomposition (SVD) and Generalized Singular Value Decomposition (GSVD). In N.J. Salkind (Ed.): \emph{Encyclopedia of Measurement and Statistics}.Thousand Oaks (CA): Sage. pp. 907-912.\cr
Abdi, H. & Williams, L.J. (2010). Barycentric discriminant analysis (BADIA). In N.J. Salkind, D.M., Dougherty, & B. Frey (Eds.): \emph{Encyclopedia of Research Design}. Thousand Oaks (CA): Sage. pp. 64-75. \cr
Abdi, H. (2007). Discriminant correspondence analysis. In N.J. Salkind (Ed.): \emph{Encyclopedia of Measurement and Statistics}. Thousand Oaks (CA): Sage. pp. 270-275.
}
\keyword{ package }
\keyword{ multivariate }
\seealso{
\code{\link{tepBADA}}, \code{\link{tepDICA}}
}
\examples{
#For more examples, see each individual function (as noted above).
}
