% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastEucCalc.R
\name{fastEucCalc}
\alias{fastEucCalc}
\title{fastEucCalc}
\usage{
fastEucCalc(x, c)
}
\arguments{
\item{x}{a set of points.}

\item{c}{a set of centers.}
}
\value{
\item{a distance matrix}{Euclidean distances of each point to each
center are returned.}
}
\description{
Fast Euclidean distance calculations.
}
\details{
This function is especially useful for discriminant analyses. The distance
from each point in \code{x} to each point in \code{c} is computed and
returned as a \code{nrow(x)} x \code{nrow(c)} matrix.
}
\author{
Hervé Abdi, Derek Beaton
}
\keyword{misc}
