\name{print.tepAssign}
\alias{print.tepAssign}
\title{Print assignment results}
\description{
	Print assignment results.
}
\usage{
	\method{print}{tepAssign}(x,\dots)
}
\arguments{
  \item{x}{an list that contains items to make into the tepAssign class.}
  \item{\dots}{inherited/passed arguments for S3 print method(s).}  
 }
\author{Derek Beaton, Cherise Chin-Fatt}
\keyword{print}
