\name{print.tepBADA}
\alias{print.tepBADA}
\title{Print tepBADA results}
\description{
	Print tepBADA results.
}
\usage{
	\method{print}{tepBADA}(x,\dots)
}
\arguments{
  \item{x}{an list that contains items to make into the tepBADA class.}
  \item{\dots}{inherited/passed arguments for S3 print method(s).}  
 }
\author{Derek Beaton, Cherise Chin-Fatt}
\keyword{print}
