\name{print.tepDICA}
\alias{print.tepDICA}
\title{Print tepDICA results}
\description{
	Print tepDICA results.
}
\usage{
	\method{print}{tepDICA}(x,\dots)
}
\arguments{
  \item{x}{an list that contains items to make into the tepDICA class.}
  \item{\dots}{inherited/passed arguments for S3 print method(s).}  
 }
\author{Derek Beaton, Cherise Chin-Fatt}
\keyword{print}
