\name{print.tepGraphs}
\alias{print.tepGraphs}
\title{Print tepGraphs results}
\description{
	Print tepGraphs results.
}
\usage{
	\method{print}{tepGraphs}(x,\dots)
}
\arguments{
  \item{x}{an list that contains items to make into the tepGraphs class.}
  \item{\dots}{inherited/passed arguments for S3 print method(s).}  
 }
\author{Derek Beaton, Cherise Chin-Fatt}
\keyword{print}
