\name{print.tepPLS}
\alias{print.tepPLS}
\title{Print tepPLS results}
\description{
	Print tepPLS results.
}
\usage{
	\method{print}{tepPLS}(x,\dots)
}
\arguments{
  \item{x}{an list that contains items to make into the tepPLS class.}
  \item{\dots}{inherited/passed arguments for S3 print method(s).}  
 }
\author{Derek Beaton, Cherise Chin-Fatt}
\keyword{print}
