% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.tepPLSCA.R
\name{print.tepPLSCA}
\alias{print.tepPLSCA}
\title{Print tepPLSCA results}
\usage{
\method{print}{tepPLSCA}(x, ...)
}
\arguments{
\item{x}{an list that contains items to make into the tepPLSCA class.}

\item{\dots}{inherited/passed arguments for S3 print method(s).}
}
\description{
Print tepPLSCA results.
}
\author{
Derek Beaton, Cherise Chin-Fatt
}
\keyword{print}
