% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.texpoOutput.R
\name{print.texpoOutput}
\alias{print.texpoOutput}
\title{Print TExPosition results}
\usage{
\method{print}{texpoOutput}(x, ...)
}
\arguments{
\item{x}{an list that contains items to make into the texpoOutput class.}

\item{\dots}{inherited/passed arguments for S3 print method(s).}
}
\description{
Print TExPosition results.
}
\author{
Derek Beaton, Cherise Chin-Fatt
}
\keyword{print}
